+++ src/main/java/org/pyneo/maps/utils/Ut.java
@@ -254,7 +254,7 @@ public class Ut implements Constants {
 		byte b[] = new byte[4];
 
 		if (in.read(b) > 0)
-			res = (((int)(b[0] & 0xFF)) << 24) +
+			res = ((b[0] & 0xFF) << 24) +
 				+((b[1] & 0xFF) << 16) +
 				+((b[2] & 0xFF) << 8) +
 				+(b[3] & 0xFF);
